requirements.txt
simplematrixbotlib
python-dotenv